+++ src/collection/domain/models/CollectionFeaturedItem.ts
@@ -2,7 +2,6 @@ export type CollectionFeaturedItem = {
   id: string
   type: 'custom'
   order: number
-  title: string
   content: string
   imageUrl?: string
 }
