+++ src/sections/collection-featured-items/featured-items-form/FeaturedItemsForm.module.scss
@@ -9,13 +9,3 @@
   display: flex;
   justify-content: flex-end;
 }
-
-.show-data-checkbox-wrapper {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-
-  :global .form-check {
-    margin-bottom: 0;
-  }
-}
