+++ src/sections/collection-featured-items/featured-items-form/featured-item-field/ContentField.tsx
@@ -1,0 +1,58 @@
+import { useTranslation } from 'react-i18next'
+import { Controller, UseControllerProps, useFormContext } from 'react-hook-form'
+import { Col, Form, RichTextEditor } from '@iqss/dataverse-design-system'
+import styles from './FeaturedItemField.module.scss'
+
+interface ContentFieldProps {
+  itemIndex: number
+}
+
+const ContentField = ({ itemIndex }: ContentFieldProps) => {
+  const { control } = useFormContext()
+  const { t } = useTranslation('collectionFeaturedItems')
+
+  const rules: UseControllerProps['rules'] = {
+    required: t('form.content.required'),
+    validate: (value: string) => {
+      const content = value.replace(/<p[^>]*>|<\/p>/g, '').trim()
+      const isEmptyTag = content === ''
+
+      if (isEmptyTag || value === '') {
+        return t('form.content.required')
+      }
+      return true
+    }
+  }
+
+  return (
+    <Form.Group as={Col} className={styles['form-group-content']}>
+      <Form.Group.Label required={true} id={`featuredItems.${itemIndex}.content`}>
+        {t('form.content.label')}
+      </Form.Group.Label>
+
+      <Controller
+        name={`featuredItems.${itemIndex}.content`}
+        control={control}
+        rules={rules}
+        render={({ field: { onChange, ref, value }, fieldState: { invalid, error } }) => {
+          return (
+            <Col>
+              <RichTextEditor
+                initialValue={value as string}
+                editorContentAriaLabelledBy={`featuredItems.${itemIndex}.content`}
+                onChange={onChange}
+                invalid={invalid}
+                ariaRequired
+                ref={ref}
+              />
+
+              {invalid && <div className={styles['error-msg']}>{error?.message}</div>}
+            </Col>
+          )
+        }}
+      />
+    </Form.Group>
+  )
+}
+
+export default ContentField
