+++ src/sections/collection-featured-items/types.ts
@@ -5,7 +5,6 @@ export type FeaturedItemsFormData = {
 }
 
 export type FeaturedItemField = {
-  title: string
   content: string
   image: File | string | null
   itemId?: CollectionFeaturedItem['id']
